requirements.txt
-f file:///home/tin/.pip_packages
Django==1.6.1
Pillow==2.0.0
South==0.8.1
argparse==1.2.1
cssselect==0.8
django-appconf==0.6
django-autoslug==1.6.1
django-bootstrap-form==3.1
django-cities-light==2.0.7
django-extensions==1.1.1
django-forms-bootstrap==3.0.0
django-jsonfield==0.8.12
django-model-utils==1.4.0
django-treebeard==2.0b1
django-user-accounts==1.0b8
djangorestframework==2.3.12
eventlog==0.6.2
factory-boy==2.3.1
ipdb==0.7
ipython==0.13.2
lxml==3.2.1
metron==1.1
pinax-theme-bootstrap==5.0.3
pinax-theme-bootstrap-account==1.0b2
pinax-utils==1.0b1.dev3
progressbar==2.2
pyquery==1.2.4
pytz==2012j
six==1.5.2
wsgiref==0.1.2
